% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runIFCshinyApp.R
\name{runIFCshinyApp}
\alias{runIFCshinyApp}
\title{Run an IFCshiny Application}
\usage{
runIFCshinyApp(
  app = "IFCshiny",
  host = "127.0.0.1",
  launch.browser = NULL,
  local = TRUE,
  ...
)
}
\arguments{
\item{app}{an app name from IFCshiny package. Default is "IFCshiny".}

\item{host}{IPv4 address that the application should listen on. Defaults is "127.0.0.1".}

\item{launch.browser}{whether to run the app and launch default system browser. Default is NULL.\cr
If set to NULL and in Rstudio, the app will be run using and displayed in Rstudio's viewer.\cr
If set to TRUE, the app will be run and the default system browser will be launched.\cr
If set to FALSE, the app will be run but no browser will be launched.\cr}

\item{local}{TRUE, FALSE or an environment, determining where the application files should be sourced. Default is TRUE.}

\item{...}{Other arguments to be passed. 
Arguments will be first passed to runApp function.\cr
Then, remaining named arguments will be passed as global variable to 'app'.}
}
\value{
it invisibly returns an object representing the app. See ?shiny::shinyApp.
}
\description{
Function to run IFCshiny application
}
\details{
it will source application files in environment set by `local` and then use shiny::shinyApp
}
\examples{
if(interactive()) runIFCshinyApp()
}
