% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runIFCshinyApp.R
\name{runIFCshinyAppOnce}
\alias{runIFCshinyAppOnce}
\title{Run Only One Instance of IFCshiny Application}
\usage{
runIFCshinyAppOnce(options = list())
}
\arguments{
\item{options}{options to pass to shiny::shinyApp.}
}
\value{
it invisibly returns an object representing the app. See ?shiny::shinyApp.
}
\description{
Function to run IFCshiny application
}
